Added documents migration and improved seeds (#418)
* Added lock documents from v1

* Disabled actions in UI if documents locked

* Added remove delete action if locked

* Added migration instead of seed

* Updated seed

* Deleted seeds

* Fixed do not update for registration if locked

* wip

* Improved logging tool

* Added progress log to update-cached-outstanding-balance migration

* Delete 1733748412-update-deleted-order-numbers.ts

* Moved 1733994455-balance-item-status-open

* Delete 1733319079-fill-paying-organization-ids.ts

* Deleted old fix seeds for keeo

* Delete 1728928974-update-cached-outstanding-balance-from-items.ts

* Added progress log to update cached-balances seed

* Added SeedTools

* Added progress logging to duplicate-member-numbers

* Added progress logging to schedule-stock-updates

* Added progress logging to groups-registration-periods

* Improved force-save-members speed

* Added progress logging to duplicate member numbers

* Added progress logging for update orders

* Added progress logging to records-configuration migration

* Moved logging tools

* Replaced charset migration with ts script

* Improved user-permissions migration

+++ backend/shared/models/src/models/DocumentTemplate.ts
@@ -1,7 +1,7 @@
 import { column } from '@simonbackx/simple-database';
 import { isSimpleError, isSimpleErrors, SimpleError } from '@simonbackx/simple-errors';
 import { QueueHandler } from '@stamhoofd/queues';
-import type { Parent, RecordAnswer} from '@stamhoofd/structures';
+import type { Parent, RecordAnswer } from '@stamhoofd/structures';
 import { BalanceItemStatus, DocumentData, DocumentPrivateSettings, DocumentSettings, DocumentStatus, DocumentTemplatePrivate, GroupType, NationalRegisterNumberOptOut, RecordAddressAnswer, RecordAnswerDecoder, RecordDateAnswer, RecordPriceAnswer, RecordSettings, RecordTextAnswer, RecordType } from '@stamhoofd/structures';
 import { Formatter, Sorter } from '@stamhoofd/utility';
 import { v4 as uuidv4 } from 'uuid';
@@ -36,6 +36,9 @@ export class DocumentTemplate extends QueryableModel {
 
     @column({ type: 'string' })
     status = DocumentStatus.Draft;
+
+    @column({ type: 'boolean' })
+    isLocked = false;
 
     @column({ type: 'boolean' })
     updatesEnabled = true;
@@ -86,6 +89,18 @@ export class DocumentTemplate extends QueryableModel {
         skipUpdate: true,
     })
     updatedAt: Date;
+
+    override save(forceSave = false): Promise<boolean> {
+        if (!forceSave && this.isLocked) {
+            throw new SimpleError({
+                code: 'locked',
+                message: 'Document template is locked',
+                human: $t(`Dit document is vergrendeld en kan niet meer gewijzigd worden.`),
+            });
+        }
+
+        return super.save();
+    }
 
     getPrivateStructure() {
         return DocumentTemplatePrivate.create(this);
@@ -363,8 +378,7 @@ export class DocumentTemplate extends QueryableModel {
                                 found = true;
                                 fieldAnswers.set(field.id, clone);
                                 break;
-                            }
-                            else {
+                            } else {
                                 console.warn('Found type mismatch for default data: ' + linkedToMemberAnswerSettingsId + ' - ' + field.id);
                             }
                         }
@@ -412,8 +426,7 @@ export class DocumentTemplate extends QueryableModel {
             for (const answer of fieldAnswers.values()) {
                 try {
                     answer.validate();
-                }
-                catch (e) {
+                } catch (e) {
                     missingData = true;
 
                     console.log('Missing data because of validation error', e, answer, answer.settings);
@@ -453,6 +466,14 @@ export class DocumentTemplate extends QueryableModel {
     }
 
     async updateForRegistration(registration: RegistrationWithMember, existingDocuments?: Document[]): Promise<Document[]> {
+        if (this.isLocked) {
+            throw new SimpleError({
+                code: 'locked',
+                message: 'Document template is locked',
+                human: $t(`Dit document is vergrendeld en kan niet meer gewijzigd worden.`),
+            });
+        }
+
         existingDocuments = existingDocuments !== undefined ? existingDocuments : await Document.where({ templateId: this.id, registrationId: registration.id }, { limit: 5 });
 
         if (!this.checkRegistrationIncluded(registration)) {
@@ -549,8 +570,7 @@ export class DocumentTemplate extends QueryableModel {
                 if (age > this.settings.maxAge) {
                     return false;
                 }
-            }
-            else {
+            } else {
                 console.warn('Missing registration.startDate in fieldAnswers when checking maxAge');
             }
         }
@@ -581,11 +601,20 @@ export class DocumentTemplate extends QueryableModel {
     async buildAll({ generateNumbers = false } = {}) {
         QueueHandler.abort('documents-build-all/' + this.id);
         return await QueueHandler.schedule('documents-build-all/' + this.id, async ({ abort }) => {
+            if (this.isLocked) {
+                return await Document.where({ templateId: this.id });
+            }
+
             if (!this.updatesEnabled) {
                 // Check status
                 const documents = await Document.where({ templateId: this.id });
                 for (const document of documents) {
                     abort.throwIfAborted();
+
+                    if (document.isLocked) {
+                        continue;
+                    }
+
                     await this.updateAnswers(document); // Only update global data
                     if (document.status === DocumentStatus.Draft || document.status === DocumentStatus.Published) {
                         document.status = this.status;
@@ -597,6 +626,11 @@ export class DocumentTemplate extends QueryableModel {
                 if (generateNumbers) {
                     for (const document of documents) {
                         abort.throwIfAborted();
+
+                        if (document.isLocked) {
+                            continue;
+                        }
+
                         if (document.number === null && document.status === DocumentStatus.Published) {
                             document.number = this.nextNumberForDocuments(documents);
                             await document.save();
@@ -627,11 +661,15 @@ export class DocumentTemplate extends QueryableModel {
             const documents = await Document.where({ templateId: this.id });
             for (const document of documents) {
                 abort.throwIfAborted();
+
+                if (document.isLocked) {
+                    continue;
+                }
+
                 if (!documentSet.has(document.id)) {
                     if (document.number === null) {
                         await document.delete();
-                    }
-                    else {
+                    } else {
                         document.status = DocumentStatus.Deleted;
                         await document.save();
                     }
@@ -644,6 +682,11 @@ export class DocumentTemplate extends QueryableModel {
             if (generateNumbers) {
                 for (const document of allDocuments) {
                     abort.throwIfAborted();
+
+                    if (document.isLocked) {
+                        continue;
+                    }
+
                     if (document.number === null && document.status === DocumentStatus.Published) {
                         document.number = this.nextNumberForDocuments(allDocuments);
                         await document.save();
@@ -733,8 +776,7 @@ export class DocumentTemplate extends QueryableModel {
             const context = await this.buildContext(organization);
             const renderedHtml = await render(this.privateSettings.templateDefinition.xmlExport, context);
             return renderedHtml;
-        }
-        catch (e) {
+        } catch (e) {
             if (isSimpleError(e) || isSimpleErrors(e)) {
                 throw e;
             }
@@ -753,8 +795,7 @@ export class DocumentTemplate extends QueryableModel {
             answer.settings = field;
             try {
                 answer.validate();
-            }
-            catch (e) {
+            } catch (e) {
                 // Invalid
                 return false;
             }
@@ -774,8 +815,7 @@ export class DocumentTemplate extends QueryableModel {
                 if (!existing.isReviewedAfter(addAnswer)) {
                     newAnswers.set(addAnswer.settings.id, addAnswer);
                 }
-            }
-            else {
+            } else {
                 newAnswers.set(addAnswer.settings.id, addAnswer);
             }
         }
@@ -786,8 +826,7 @@ export class DocumentTemplate extends QueryableModel {
         if (document.status !== DocumentStatus.Deleted) {
             if (!complete) {
                 document.status = DocumentStatus.MissingData;
-            }
-            else {
+            } else {
                 if (document.status === DocumentStatus.MissingData) {
                     document.status = this.status;
                 }
